+++ packages/features/sales-escrows/index.ts
@@ -1,4 +1,4 @@
-export * from './modals/ConfirmSalesActionModal';
+export * from './modals/ConfirmEndSaleModal';
 export * from './modals/StartAuctionModal';
 export * from './modals/StartEscrowModal';
 export * from './modals/TransferTokens';
