--- packages/features/sales-escrows/modals/ConfirmSalesActionModal.tsx
@@ -1,260 +1,0 @@
-import React, { useEffect, useState } from 'react';
-import Slide from '@mui/material/Slide';
-import { TransitionProps } from '@mui/material/transitions';
-import { AuthContext } from '@dapp/features-authentication';
-import { LoadingContainer } from '@dapp/features-components';
-import { useSnackbar } from 'notistack';
-import { Container, Flex, Modal, Button, HR } from '@origyn-sa/origyn-art-ui';
-import { useTokensContext } from '@dapp/features-tokens-provider';
-import { Principal } from '@dfinity/principal';
-import { EscrowReceipt, EscrowRecord } from '@dapp/common-types';
-import { useDebug } from '@dapp/features-debug-provider';
-
-const Transition = React.forwardRef(
-  (
-    props: TransitionProps & {
-      children: React.ReactElement<any, any>;
-    },
-    ref: React.Ref<unknown>,
-  ) => <Slide direction="up" ref={ref} {...props} />,
-);
-
-Transition.displayName = 'Transition';
-
-interface ConfirmSalesActionModalProps {
-  openConfirmation: boolean;
-  onClose: () => void;
-  currentToken: string;
-  action: string;
-  escrow?: EscrowRecord;
-  offer?: EscrowRecord;
-  onSaleCancelled?: () => void;
-  onProcessing?: (boolean) => void;
-}
-
-export const ConfirmSalesActionModal = ({
-  openConfirmation,
-  onClose,
-  currentToken,
-  action,
-  escrow = null,
-  offer,
-  onSaleCancelled,
-  onProcessing,
-}: ConfirmSalesActionModalProps) => {
-  const { actor, principal } = React.useContext(AuthContext);
-  const [isLoading, setIsLoading] = React.useState(false);
-  const { enqueueSnackbar } = useSnackbar() || {};
-  const { tokens } = useTokensContext();
-  const [confirmed, setConfirmed] = useState(false);
-  const debug = useDebug();
-
-  const _handleClose = async (confirm = false) => {
-    if (!confirm) {
-      onClose();
-    } else if (isLoading || !actor) {
-      return;
-    }
-    try {
-      setIsLoading(true);
-      setConfirmed(true);
-      onProcessing?.(true);
-      if (action === 'endSale') {
-        const endSaleResponse = await actor.sale_nft_origyn({
-          end_sale: currentToken,
-        });
-        if (endSaleResponse.ok) {
-          enqueueSnackbar(`You have successfully ended the sale for ${currentToken}.`, {
-            variant: 'success',
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-          });
-          onSaleCancelled();
-        }
-        enqueueSnackbar(`Error: ${endSaleResponse.err.flag_point}.`, {
-          variant: 'error',
-          anchorOrigin: {
-            vertical: 'top',
-            horizontal: 'right',
-          },
-        });
-      }
-      if (action === 'withdraw') {
-        if (!escrow) {
-          return onClose();
-        }
-        const withdrawResponse = await actor?.sale_nft_origyn({
-          withdraw: {
-            escrow: {
-              ...escrow,
-              withdraw_to: { principal },
-            },
-          },
-        });
-
-        if ('err' in withdrawResponse) {
-          enqueueSnackbar(`Error: ${withdrawResponse.err.flag_point}.`, {
-            variant: 'error',
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-          });
-        } else {
-          enqueueSnackbar('Your escrow has been successfully withdrawn.', {
-            variant: 'success',
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-          });
-        }
-      }
-
-      if (action === 'reject') {
-        if (!escrow) {
-          return onClose();
-        }
-        const rejectResponse = await actor?.sale_nft_origyn({
-          withdraw: {
-            reject: {
-              ...escrow,
-            },
-          },
-        });
-
-        if ('err' in rejectResponse) {
-          enqueueSnackbar(`Error: ${rejectResponse.err.text}.`, {
-            variant: 'error',
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-          });
-        } else {
-          enqueueSnackbar('The escrow has been rejected.', {
-            variant: 'success',
-            anchorOrigin: {
-              vertical: 'top',
-              horizontal: 'right',
-            },
-          });
-        }
-      }
-
-      const escrowReceipt: EscrowReceipt = {
-        seller: { principal: offer.seller['principal'] },
-        buyer: { principal: offer.buyer['principal'] },
-        token_id: currentToken,
-        token: {
-          ic: {
-            fee: BigInt(tokens[offer.token['ic'].symbol]?.fee ?? 200000),
-            decimals: BigInt(tokens[offer.token['ic'].symbol]?.decimals ?? 8),
-            canister: Principal.fromText(tokens[offer.token['ic'].symbol]?.canisterId),
-            standard: { Ledger: null },
-            symbol: offer.token['ic'].symbol,
-          },
-        },
-        amount: BigInt(offer.amount),
-      };
-
-      const saleReceipt = {
-        broker_id: [],
-        pricing: { instant: null },
-        escrow_receipt: [escrowReceipt],
-      };
-
-      if (action === 'acceptOffer') {
-        try {
-          const acceptOffer = await actor.market_transfer_nft_origyn({
-            token_id: currentToken,
-            sales_config: saleReceipt,
-          });
-          debug.log(acceptOffer.err);
-          if ('err' in acceptOffer) {
-            enqueueSnackbar('There has been an error in accepting the offer', {
-              variant: 'error',
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-            });
-          } else {
-            enqueueSnackbar('The offer has been accepted.', {
-              variant: 'success',
-              anchorOrigin: {
-                vertical: 'top',
-                horizontal: 'right',
-              },
-            });
-          }
-        } catch (e) {
-          debug.log(e);
-        }
-      }
-    } catch (e) {
-      debug.log(e);
-    } finally {
-      onProcessing?.(false);
-      setIsLoading(false);
-      onClose();
-    }
-  };
-
-  useEffect(() => {
-    setConfirmed(false);
-  }, [openConfirmation]);
-
-  return (
-    <Modal isOpened={openConfirmation} closeModal={() => onClose()} size="md">
-      <Container size="full" padding="48px">
-        <h2>
-          {action === 'acceptOffer'
-            ? 'Confirm Accept Offer'
-            : action === 'endSale'
-            ? 'Confirm End Sale'
-            : action === 'withdraw'
-            ? 'Confirm Escrow Withdraw'
-            : 'Confirm Escrow Rejection'}
-        </h2>
-        <br />
-        <Flex flexFlow="column">
-          {action === 'acceptOffer' ? (
-            <div>
-              Are you sure you want to accept the offer for token <b>{currentToken}</b> ?
-            </div>
-          ) : action === 'endSale' ? (
-            <div>
-              Are you sure you want to end the sale for token <b>{currentToken}</b> ?
-            </div>
-          ) : action === 'withdraw' ? (
-            <>Are you sure you want to withdraw the escrow?</>
-          ) : (
-            <>Are you sure you want to reject the escrow?</>
-          )}
-        </Flex>
-        <HR marginTop={24} marginBottom={24} />
-        <Flex flow="row" justify="flex-end" gap={16}>
-          <Flex>
-            <Button onClick={() => _handleClose(false)} disabled={confirmed}>
-              Cancel
-            </Button>
-          </Flex>
-          <Flex>
-            <Button onClick={() => _handleClose(true)} variant="contained" disabled={confirmed}>
-              Confirm
-            </Button>
-          </Flex>
-        </Flex>
-        {isLoading && (
-          <>
-            <HR marginTop={24} marginBottom={24} />
-            <LoadingContainer />
-          </>
-        )}
-      </Container>
-    </Modal>
-  );
-};
